ChainsIO | Correctly working chain streaming
I removed beloved decimal, it have to many types of zero for me.
Change hashing iternals, now it uses serrializations, yeah, it bad, but it works.

+++ Microcoin/Blockchain/ChainController/DeepTransactionsVerify.cs
@@ -33,7 +33,7 @@ namespace Microcoin.Blockchain.ChainController
         {
             // Count summary coins difference after this block
             cancellationToken.ThrowIfCancellationRequested();
-            Dictionary<string, decimal> tempCoinsCount = new Dictionary<string, decimal>();
+            Dictionary<string, double> tempCoinsCount = new Dictionary<string, double>();
             foreach (Transaction.Transaction transaction in transactions)
             {
                 if (tempCoinsCount.TryAdd(transaction.SenderPublicKey, -transaction.TransferAmount) is not true)
