+++ Microcoin/Blockchain/Mining/IRewardRule.cs
@@ -10,6 +10,6 @@ namespace Microcoin.Blockchain.Mining
     public interface IRewardRule
     {
         public bool Verify(AbstractChain contextChain, Block.Block block);
-        public decimal Calculate(AbstractChain contextChain, Block.Block block);
+        public double Calculate(AbstractChain contextChain, Block.Block block);
     }
 }
